+++ components/study-tracker/index.ts
@@ -1,8 +1,6 @@
 export { StudyTracker } from './StudyTracker';
 export { StudyWeeklyView } from './StudyWeeklyView';
-export { StudyMonthlyView } from './StudyMonthlyView';
 export { StudyTaskItem } from './StudyTaskItem';
 export { SubjectManagementModal } from './SubjectManagementModal';
 export { StudyCell } from './StudyCell';
 export { StudySubjectRow } from './StudySubjectRow';
-export { StudyFilters } from './StudyFilters';